pyproject.toml
[tool.poetry]
name = "ethereum_analysis"
version = "0.1.0"
description = "analyzing historical ethereum transactions"
authors = ["Vikram Saraph <vikram.saraph.22@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.10"
PyYAML = "^6.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
